requirements.txt
Flask
pymongo[srv]
requests
gunicorn
python-dotenv